ajout makefile et fix format / nom

+++ Libft/ft_strncmp.c
@@ -6,15 +6,17 @@
 /*   By: yseguin <yseguin@student.42perpignan.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 15:49:03 by yseguin           #+#    #+#             */
-/*   Updated: 2024/11/04 15:51:16 by yseguin          ###   ########.fr       */
+/*   Updated: 2024/11/05 21:36:40 by yseguin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
+
 ////////////////////////////////////////////////////////////////
 //func for compqre s1 and s2 with n lenth
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n && (s1[i] != '\0' && s2[i] != '\0'))
